Setup Instant HIE Core package on AWS
The AWS scripts required a different setup mechanism compared to the
minikube setup. Minikube used host name aliasing on the local machine to
allow the openhim console to connect to the openhim core without having
to know the exact url of the core.

That mechanism couldn't work for AWS as the hostnames in this case
weren't IPs but URLs provided by AWS which could not be used in the
local HOST file.

This mechanism sets up the Hapi FHIR and OpenHIM core in the first part
of the script and gets their urls. The OpenHIM Core url is then
substituted directly into the openhim console config file which is then
used to create a configmap. Therefore the configmap contains a reference
to the OpenHIM Core's api endpoint - which shouldn't change.

OHM-972