use labels for identifying packages instead of namespaces
Namespaces were being used for isolating the different packages (the core and healthworkforce packages) but this has been changed as both the packages share a database (a replica set). A replica set in a namespace can be accessed within the cluster in its namespace and outside the cluster but accessing it from another namespace (virtual cluster in the cluster) is not possible. The packages have been moved to the default namespace and labels have been added to the components of the packages enable us to identify the components of a package

OHIE-395