change the mock server to be a vanilla server
The mock server for the gofr and ihris services for the mcsd mediator external requests was previously an express server and has been changed to be a vanilla server with no third party dependencies. k8s does not have the build feature like docker-compose for building an image. The image has to be prebuilt. Building an image and uploading it to dockerhub where it will be pulled, involves a lot work and as well the mock services are only used now and will be thrown away once we have gofr and ihris endpoints. To deploy the mock services, we make use of jembi/config-importer image. The files for mock services are mounted to the container and then ran. With a vanilla server there arent third party party libraries that the server uses and this is perfect as we do not have to install anything beacuse the container's file system is a readonly system. With an express server there are dependencies which couldn't be installed as there is no write access

OHIE-395