update readme with new instructions
a new option for the command for setting up the services has been added called the 'init' script. This is the option in which channel configurations and replica set intialisation will be done. The 'up' option is now only for starting up the services and does not have any configuratons.

OHIE-395