Updated mcsd mock service container to different version
The current latest version which made use of the samples provided caused unforseen issues with the transaction timing out due to the size of the transaction.
The slim version of the mock services reduced the size of the sample resources making the request smaller and successfully succeed.

This needs futher investigation into what is causing the failure

OHIE-395