Specify NODE_PATH to globally installed modules
Node scripts won't have easy access to the globally installed modules
without the NODE_PATH set.
This is needed for the Axios module mostly. This is a light weight image.
But we have added Axios as it is a useful module making http/s requests.
Node Scripts will now be able to access this moduile using `require('axios')`
instead of having to reference the Global NPM storage location.

MENC-21