Convert instant.sh to isntant.ts
As the work that instant.sh needs to do get more complex, we need a language that is easier to develop in. Converting this to js allows us to more easily handle complex commandline options and will eventually allow us to handle complex structure like dependency trees easier than in bash.

OHIE-420