update the version of the mock service
New features have been added to the mock service.Post requests can now be made to create and delete requests to delete the created resources. THis is to allow the testing to be easy as we will know exactly what will be sent to the FHIR server (for assertions and cleanup). THe resources created can then be deleted after testing (from FHIR and the mock services). The mock service is used for other things and deleting a created resource will reset it to return the resources it normally returns

OHIE-423