Make the client opencr config more generic
OpenCR will require config per use case as it requires system identifies
of the use case provided on start up config.
Custom packages will, for the moment, need to append their config to the
default config of the client package. Therefore the client package
config should be as minimal and generic as possible.

The default systems.internalid value provided is a place holder to
allow the system to start up

DEb-5