Add the replica-set command to core dev mongo start up
In the development docker-compose script the mongo instance was being
started up with the default start command therefore it was not set up to
be part of a replica set. This caused the mongo instance to crash an
restart when attempting to make it a replica set. The instance would no
be part of the set.

DEB-5