Simplify the DHIS2 testing
The tests here configured the DHIS2 instance to allow for testing a
scenario - unfortunately it is very difficutl to remove config from
DHIS2 once it have been imported.

DHIS2 configuration is quite difficult to get right and having an
existing config file in the program beforehand increases the difficulty.

Therefore the current test will just check the DHIS2 is running and that
an authenticated user can get system information

DSC19-177