Better specify node version so that image builds
The older version of debina that we were using the node:lts tag is now not supported. This lead to some of the tool we are installing in the docekrfile to not install correctly. This fix ensures that the latest supported version of debian is used in the underlying image.