Prevent mongo rep set from using more ram than a system has
The mongos are now restricted to 512Mb RAM for their cache. Before this change server would crash eventually as mongo by default tries to use 50% of system RAM for it's cache - when running 3 monos as a rep set on a single server that is 150% RAM... whcih is never going to work.

DIO-22