Fix mcsd tests - broken trigger
The tests attempted to trigger the polling channel with a GET request to
the public API - this is not possible for polling channels. Polling
channels can be triggered via an API call to the mediator API using the
channel ID. This is a lot of work to get therefore I have switched the
channel to a regualr HTTP channel that will allow easier access.

DIO-52