Add support for setting the OpenHIM root password via env var
The OPENHIM_ROOT_PASSWORD env var, when supplied to the instant orchestration container will create and set the root password in the default OpenHIM config that get uploaded once the OpenHIM is up and running.

To test this you will need to add the line '-e "OPENHIM_ROOT_PASSWORD=test" \' to line 44 of the ./deploy.sh script. This is until we support sending in env vars from the GO CLI.